refactor: split expenses-structure widget into smaller components
Extract the monolithic expenses-structure.vue (519 lines) into a folder
with focused components: data composable, donut chart, excluded
categories popover, and an orchestrator index.vue
