+++ packages/frontend/src/components/widgets/widget-registry.ts
@@ -49,7 +49,7 @@ export const WIDGET_REGISTRY: Record<string, WidgetDefinition> = {
       { colSpan: 1, rowSpan: 1, label: '1×1' },
       { colSpan: 2, rowSpan: 1, label: '2×1' },
     ],
-    component: () => import('@/components/widgets/expenses-structure.vue'),
+    component: () => import('@/components/widgets/expenses-structure/index.vue'),
     needsPeriod: true,
   },
   'latest-records': {
